Wrap Implementations, split out font-descriptor methods
